requirements.txt
python-pptx>=1.0.0,<1.1.0
Pillow==10.4.0
google-generativeai==0.7.2
tenacity==8.3.0
rapidfuzz==3.9.6
orjson==3.10.7
typer==0.12.3
rich==13.7.1
python-dateutil==2.9.0.post0
uvloop==0.19.0; platform_system != 'Windows'

